Extract groups of date-time functions into own classes
