+++ h2/src/main/org/h2/expression/BinaryOperation.java
@@ -7,8 +7,7 @@ package org.h2.expression;
 
 import org.h2.engine.Session;
 import org.h2.expression.IntervalOperation.IntervalOpType;
-import org.h2.expression.function.DateTimeFunctions;
-import org.h2.expression.function.Function;
+import org.h2.expression.function.DateTimeFunction;
 import org.h2.message.DbException;
 import org.h2.value.DataType;
 import org.h2.value.TypeInfo;
@@ -303,19 +302,16 @@ public class BinaryOperation extends Operation2 {
             switch (l) {
             case Value.INTEGER:
                 // Oracle date add
-                return Function.getFunctionWithArgs(Function.DATEADD,
-                        ValueExpression.get(ValueInteger.get(DateTimeFunctions.DAY)), left, right).optimize(session);
+                return new DateTimeFunction(left, right, DateTimeFunction.DATEADD, DateTimeFunction.DAY)
+                        .optimize(session);
             case Value.NUMERIC:
             case Value.REAL:
             case Value.DOUBLE:
                 // Oracle date add
-                return Function
-                        .getFunctionWithArgs(Function.DATEADD,
-                                ValueExpression.get(ValueInteger.get(DateTimeFunctions.SECOND)),
-                                new BinaryOperation(OpType.MULTIPLY,
-                                        ValueExpression.get(ValueInteger.get(60 * 60 * 24)), left),
-                                right)
-                        .optimize(session);
+                return new DateTimeFunction(
+                        new BinaryOperation(OpType.MULTIPLY, ValueExpression.get(ValueInteger.get(60 * 60 * 24)), //
+                                left),
+                        right, DateTimeFunction.DATEADD, DateTimeFunction.SECOND).optimize(session);
             }
             break;
         }
@@ -330,10 +326,8 @@ public class BinaryOperation extends Operation2 {
                         throw getUnexpectedForcedTypeException();
                     }
                     // Oracle date subtract
-                    return Function.getFunctionWithArgs(Function.DATEADD,
-                            ValueExpression.get(ValueInteger.get(DateTimeFunctions.DAY)), //
-                            new UnaryOperation(right), //
-                            left).optimize(session);
+                    return new DateTimeFunction(new UnaryOperation(right), left, DateTimeFunction.DATEADD,
+                            DateTimeFunction.DAY).optimize(session);
                 }
                 case Value.NUMERIC:
                 case Value.REAL:
@@ -342,11 +336,10 @@ public class BinaryOperation extends Operation2 {
                         throw getUnexpectedForcedTypeException();
                     }
                     // Oracle date subtract
-                    return Function.getFunctionWithArgs(Function.DATEADD,
-                                ValueExpression.get(ValueInteger.get(DateTimeFunctions.SECOND)),
-                                new UnaryOperation(new BinaryOperation(OpType.MULTIPLY, //
-                                        ValueExpression.get(ValueInteger.get(60 * 60 * 24)), right)), //
-                                left).optimize(session);
+                    return new DateTimeFunction(
+                            new BinaryOperation(OpType.MULTIPLY, ValueExpression.get(ValueInteger.get(-60 * 60 * 24)),
+                                    right),
+                            left, DateTimeFunction.DATEADD, DateTimeFunction.SECOND).optimize(session);
                 }
                 case Value.TIME:
                 case Value.TIME_TZ:
