+++ h2/src/main/org/h2/expression/function/MathFunction1.java
@@ -116,22 +116,6 @@ public class MathFunction1 extends Operation1 implements NamedExpression {
             "LOG10", "LN", "EXP", "SQRT", "DEGREES", "RADIANS" //
     };
 
-    /**
-     * Returns whether specified function is known by this class.
-     *
-     * @param upperName
-     *            the name of the function in upper case
-     * @return {@code true} if it exists
-     */
-    public static boolean exists(String upperName) {
-        for (String n : NAMES) {
-            if (n.equals(upperName)) {
-                return true;
-            }
-        }
-        return false;
-    }
-
     private final int function;
 
     public MathFunction1(Expression arg, int function) {
